Fix dtype mismatch: use bfloat16 instead of fp16
fp16 causes 'Half and Float' errors with batch_size>1 in NeMo.
bfloat16 handles mixed precision properly and is recommended by
NVIDIA for NeMo inference.

+++ backend/src/cassandra_yt_mcp/services/local_transcriber.py
@@ -71,9 +71,9 @@ class LocalTranscriber:
                 model = nemo_asr.models.ASRModel.from_pretrained(
                     model_name=_PARAKEET_MODEL, map_location=self._device,
                 )
-                # fp16 halves CPU RAM for weights and speeds up inference
+                # bfloat16 reduces VRAM and speeds up inference on modern GPUs
                 if self._device != "cpu":
-                    model = model.to(dtype=torch.float16)
+                    model = model.to(dtype=torch.bfloat16)
             model.change_attention_model(
                 self_attention_model="rel_pos_local_attn",
                 att_context_size=[256, 256],
